Edited files

+++ custom-recipes/write-data-partitions-to-excel-sheets/recipe.py
@@ -121,15 +121,22 @@ folder_info = output_folder.get_info()
 def write_partitions():
     path = os.path.join(folder_info['path'], excel_name)
     writer = pd.ExcelWriter(path)
+     
     if partitioning_columns:
         i=0
         for dframe in dfs:
             dframe =dframe.applymap(str)
             if use_partition_value_for_sheetname:
-                dframe.to_excel(writer, sheet_name=final_sheet_names[i], startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)
+                if use_existing_file:
+                    dframe.to_excel(writer, sheet_name=final_sheet_names[i], mode='a', startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)
+                else:
+                    dframe.to_excel(writer, sheet_name=final_sheet_names[i], startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)
             else:
                 sheet_name_1 = "Sheet" + str(i+1)
-                dframe.to_excel(writer, sheet_name=sheet_name_1, startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)            
+                if use_existing_file:    
+                    dframe.to_excel(writer, sheet_name=sheet_name_1, mode='a', startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)            
+                else:  
+                    dframe.to_excel(writer, sheet_name=sheet_name_1, startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)            
             i=i+1
         writer.save()
 
@@ -141,7 +148,10 @@ def write_partitions():
         else:
            df_frame = input_data_df.copy()
         dframe =df_frame.applymap(str)
-        dframe.to_excel(writer, sheet_name=sheet_name, startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)
+        if use_existing_file:
+            dframe.to_excel(writer, sheet_name=sheet_name, mode='a', startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)
+        else:
+            dframe.to_excel(writer, sheet_name=sheet_name, startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)    
         writer.save()
 
 
@@ -150,9 +160,13 @@ def write_partitions_timestamp():
     # get current timestamp
     current_time = datetime.datetime.now()
     current_time = current_time.strftime("%m-%d-%Y-%H-%M-%S")
-    excel_name_1 = f"{excel_name}_{current_time}.xlsx"
-
-    path = os.path.join(folder_info['path'], excel_name_1)
+    
+    if use_existing_file:
+        excel_name_1 = f"{existing_file}_{current_time}.xlsx"
+        path = os.path.join(folder_info['path'], excel_name_1)
+    else:
+        excel_name_1 = f"{excel_name}_{current_time}.xlsx"
+        path = os.path.join(folder_info['path'], excel_name_1)
     writer = pd.ExcelWriter(path)
     
     if partitioning_columns:
@@ -160,10 +174,16 @@ def write_partitions_timestamp():
         for dframe in dfs:
             dframe =dframe.applymap(str)
             if use_partition_value_for_sheetname:
-                dframe.to_excel(writer, sheet_name=final_sheet_names[i], startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)
+                if use_existing_file:
+                    dframe.to_excel(writer, sheet_name=final_sheet_names[i], mode='a', startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)
+                else:
+                    dframe.to_excel(writer, sheet_name=final_sheet_names[i], startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)
             else:
                 sheet_name_1 = "Sheet" + str(i+1)
-                dframe.to_excel(writer, sheet_name=sheet_name_1, startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)            
+                if use_existing_file:    
+                    dframe.to_excel(writer, sheet_name=sheet_name_1, mode='a', startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)            
+                else:  
+                    dframe.to_excel(writer, sheet_name=sheet_name_1, startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)            
             i=i+1
         writer.save()
 
@@ -176,7 +196,10 @@ def write_partitions_timestamp():
         else:
            df_frame = input_data_df.copy()
         dframe =df_frame.applymap(str)
-        dframe.to_excel(writer, sheet_name=sheet_name, startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)
+        if use_existing_file:
+            dframe.to_excel(writer, sheet_name=sheet_name, mode='a', startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)
+        else:
+            dframe.to_excel(writer, sheet_name=sheet_name, startrow=start_row, startcol=start_col, encoding='utf-8', index = None, header = True)
         writer.save()
 
 # function calls to write data
